Add AuthController
Routes for logging in and out with keycloak are added. The assent API
generates a url and session state. The browser is redirect to this url,
and the session state is saved in the session storage for later
retrieval. Once the user is redirected back to the callback url the
session state is used again to get the user parameters after a succesful
login attempt. These parameters are used to ensure a User in the
database (using the upsert capabilities). Finally the controller hooks
into the `phx.gen.auth` system with the `UserAuth.log_in_user/2`
function.
Logging out currently only removes the local cookies. Maybe there is a
better way to also redirect to the OAuth/OIDC provider to log out there
too?