Strip generated code by phx.gen.auth
Most of the generated code is removed again. Everything related to
passwords and cookies can safely be removed again, as this will be
handled by the OAuth provider.
The UserToken also has its 'context' removed, as that will always be
"session" from now on (previously it could also contain tokens for
password resets or account confirmation).
The links for logging in and out are broken now, but we'll fix that
later on.
Also notice that we have removed the dependencies that were added by
`phx.gen.auth`, as they are not required anymore.
The migrations are also simplified, and now have no reference anymore to
a password, a confirmation timestamp, a token context and a token
send_to attribute.